Extract endpoint
