+++ lib/monobank/personal/statement.rb
@@ -1,8 +1,9 @@
+require 'monobank/endpoint'
 require 'monobank/resources/personal/statement'
 
 module Monobank
   module Personal
-    class Statement
+    class Statement < Endpoint
       ENDPOINT = '/personal/statement'.freeze
 
       def initialize(token:, account_id:, from:, to:)
@@ -12,27 +13,14 @@ module Monobank
         @to = to
       end
 
-      def call
-        attributes = connection.get(pathname, options)
-        define_resources(attributes)
-      end
-
       private
 
-      attr_reader :token, :account_id, :from, :to
+      attr_reader :account_id, :from, :to
 
       def pathname
         path = "#{ENDPOINT}/#{account_id}/#{from}"
         path = "#{path}/#{to}" if to
         path
-      end
-
-      def options
-        { headers: { "X-Token" => token.to_s } }
-      end
-
-      def connection
-        @connection ||= Connection.new
       end
 
       def define_resources(attributes)
