+++ lib/monobank/resources/bank/currency.rb
@@ -1,19 +1,10 @@
+require 'monobank/resources/base'
+
 module Monobank
   module Resources
     module Bank
-      class Currency
-        def initialize(attributes)
-          @attributes = {}
-
-          attributes.each do |key, value|
-            @attributes[method_name(key)] = value
-            self.class.define_method(method_name(key)) { value }
-          end
-        end
-
-        def method_name(key)
-          key.gsub(/(.)([A-Z])/,'\1_\2').downcase
-        end
+      class Currency < Base
+        define_fields %w[currency_code_a currency_code_b date rate_sell rate_buy rate_cross]
       end
     end
   end
