+++ lib/monobank/resources/personal/accounts.rb
@@ -1,20 +1,10 @@
+require 'monobank/resources/base'
+
 module Monobank
   module Resources
     module Personal
-      class Accounts
-        def initialize(attributes)
-
-          @attributes = {}
-
-          attributes.each do |key, value|
-            @attributes[method_name(key)] = value
-            self.class.define_method(method_name(key)) { value }
-          end
-        end
-
-        def method_name(key)
-          key.gsub(/(.)([A-Z])/,'\1_\2').downcase
-        end
+      class Accounts < Base
+        define_fields %w[id balance credit_limit currency_code cashback_type]
       end
     end
   end
